Make assertion into hook
